Redirect output to /dev/null
When you run a command with nohup (meaning "no hangup") you get a file in your current working directory containing any output that would have went to the terminal but is instead directed to this file since the process is running headless.  The change I added instead redirects that output to /dev/null and avoids creating a nohup.out file in the current directory.